refactor: [#15] Dashboard 관련 컴포넌트 폴더 구조 이동 및 정리
