+++ frontend/src/pages/Dashboard/boardbox.tsx
@@ -1,16 +1,20 @@
 import type { BoxtypeWithCards } from '@/types/board';
 
 import { Boardcard } from './boardcard';
-import { Boardmodal } from '../Modal/boardmodal';
-import { Button } from '../ui/button';
+import { Boardmodal } from '../../components/Modal/boardmodal';
+import { Button } from '../../components/ui/button';
 import {
   Card,
   CardAction,
   CardContent,
   CardHeader,
   CardTitle,
-} from '../ui/card';
-import { Dialog, DialogContent, DialogTrigger } from '../ui/dialog';
+} from '../../components/ui/card';
+import {
+  Dialog,
+  DialogContent,
+  DialogTrigger,
+} from '../../components/ui/dialog';
 
 type BoardboxProps = {
   box: BoxtypeWithCards;
